WTA #12: Moved some components around.
