+++ home/src/main/java/com/jacob/wakatimeapp/home/ui/HomePageNavigator.kt
@@ -1,0 +1,5 @@
+package com.jacob.wakatimeapp.home.ui
+
+interface HomePageNavigator {
+    fun toDetailsPage()
+}
